Add files via upload
This project is significant because CKD is a prevalent
disease that affects millions of people worldwide. Early
detection and accurate classification of CKD can
improve patient outcomes and help prevent the
progression of the disease. In addition, it includes the
accuracy of the model, its ability to identify key features
and patterns in the data, and its potential to improve
patient outcomes and prevent the progression of CKD.